backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pg8000
pydantic
python-multipart
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-dotenv
